fix: cr suggestions
