+++ package.json
@@ -39,7 +39,7 @@
     "mocha": "9.1.3",
     "nodemon": "2.0.15",
     "prettier": "2.5.1",
-    "supertest": "^6.1.6",
+    "supertest": "6.1.6",
     "ts-node": "10.4.0",
     "typescript": "4.5.4"
   },
