+++ src/setup.ts
@@ -3,14 +3,12 @@ import { ApolloServer } from 'apollo-server-express';
 import { ApolloServerPluginDrainHttpServer } from 'apollo-server-core';
 import * as express from 'express';
 import * as http from 'http';
-import { createConnection, getConnection } from 'typeorm';
+import { createConnection } from 'typeorm';
 import { User } from './entity/User';
 import * as fs from 'fs';
 import * as path from 'path';
-import * as crypto from 'crypto';
-import { UserInput } from './UserInput';
+import { resolvers } from './resolvers';
 
-//database setup
 export async function setupDatabase() {
   await createConnection({
     type: 'postgres',
@@ -24,37 +22,6 @@ export async function setupDatabase() {
     entities: [User],
   });
 }
-
-const resolvers = {
-  Query: {
-    hello() {
-      return 'Hello, world!';
-    },
-  },
-  Mutation: {
-    createUser: async (_: undefined, args: UserInput) => {
-      const user = new User();
-      const userRepository = getConnection().getRepository(User);
-
-      user.name = args.data.name;
-      user.email = args.data.email;
-      user.password = crypto.createHash('sha256').update(args.data.password).digest('hex');
-      user.birthDate = new Date(args.data.birthDate);
-
-      if (args.data.password.length < 6 || !/\d/.test(args.data.password) || !/[A-Za-z]/.test(args.data.password)) {
-        throw new Error('Weak password!');
-      }
-      const countUsers = await userRepository.count({ email: user.email });
-
-      if (countUsers > 0) {
-        throw new Error('This e-mail is already being used!');
-      }
-
-      await userRepository.insert(user);
-      return user;
-    },
-  },
-};
 
 export async function setupServer(port: number) {
   const app = express();
